requirements.txt
numpy==1.17.5
requests==2.26.0
Pillow==8.3.2